@@ -109,9 +116,9 @@
     "prop-types": "^15.6.2",
     "qs": "^6.5.2",
     "rand-token": "^0.4.0",
-    "react": "^16.8.6",
-    "react-bootstrap": "^0.32.1",
-    "react-dom": "^16.8.2",
+    "react": "^17.0.2",
+    "react-bootstrap": "^0.33.1",
+    "react-dom": "^17.0.2",
     "react-dom-factories": "^1.0.2",
     "react-hot-loader": "^4.7.1",
     "react-redux": "^7.0.1",
@@ -123,31 +130,39 @@
     "redux-thunk": "^2.3.0",
     "sequelize": "^5.21.11",
     "sequelize-cli": "^5.5.1",
-    "sequelize-paper-trail": "git+https://github.com/libertas-funding/sequelize-paper-trail.git#dist",
+    "sequelize-paper-trail": "^3.0.1",
     "sinon": "^11.1.1",
-    "sqlite3": "git+https://github.com/mapbox/node-sqlite3.git#593c9d",
+    "sqlite3": "5.0.8",
     "style-loader": "^0.21.0",
     "webpack": "^4.44.1",
+    "webpack-cli": "^4.9.2",
     "yup": "^0.32.9"
   },
   "optionalDependencies": {
     "fsevents": "^2.3.2"
   },
   "devDependencies": {
-    "@babel/cli": "^7.2.3",
-    "@babel/plugin-proposal-class-properties": "^7.3.3",
-    "@babel/plugin-proposal-object-rest-spread": "^7.3.2",
+    "@babel/cli": "^7.14.8",
+    "@babel/eslint-parser": "^7.17.0",
+    "@babel/eslint-plugin": "^7.17.7",
+    "@babel/plugin-proposal-class-properties": "^7.14.5",
+    "@babel/plugin-proposal-object-rest-spread": "^7.14.7",
+    "@testing-library/react": "^12.1.2",
     "axios-mock-adapter": "^1.15.0",
     "babel-plugin-css-modules-transform": "^1.6.2",
-    "babel-plugin-istanbul": "^5.1.1",
+    "babel-plugin-istanbul": "^6.0.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "enzyme": "^3.9.0",
-    "enzyme-adapter-react-16": "^1.6.0",
-    "eslint": "^6.8.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.14.3",
+    "enzyme": "^3.11.0",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.7",
+    "eslint": "^8.13.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-import-resolver-babel-plugin-root-import": "^1.1.1",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-mocha": "^10.0.4",
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-hooks": "^4.4.0",
     "factory-girl": "^5.0.2",
     "jsdom": "^16.6.0",
     "jsdom-global": "^3.0.2",
@@ -166,15 +181,20 @@
     "webpack-dev-middleware": "^3.6.0"
   },
   "engines": {
-    "node": "16.14.0",
-    "npm": "8.3.1"
+    "node": "16.15.0",
+    "npm": "8.5.5"
   },
   "overrides": {
     "@react-bootstrap/pagination": {
       "react": "$react",
       "react-dom": "$react-dom"
     },
+    "react-router-dom": {
+      "react": "$react",
+      "react-dom": "$react-dom"
+    },
     "react-bootstrap": {
+      "react": "$react",
       "react-dom": "$react-dom"
     },
     "react-test-renderer": {
