@@ -2,8 +2,10 @@
  * file: webpack.config.js
  */
 
+import SystemService from '~/services/system'
+
 const path = require('path');
-import SystemService from "~/services/system"
+
 const NODE_ENV = SystemService.get_node_env();
 const OPTIMIZATION_MODE = NODE_ENV === 'production' ? 'production' : 'development';
 
@@ -12,7 +14,7 @@ const OPTIMIZATION_MODE = NODE_ENV === 'production' ? 'production' : 'developmen
  */
 
 function build_by_entry_and_target(entry, target) {
-  let output = {
+  const output = {
     mode: OPTIMIZATION_MODE,
     entry,
     output: {
@@ -39,12 +41,12 @@ function build_by_entry_and_target(entry, target) {
           test: /\.css$/,
           use: ['style-loader', 'css-loader'],
         },
-      ]
+      ],
     },
   }
   return output;
 }
 
 export default [
-  build_by_entry_and_target('./client/index.jsx', 'bundle.js'),
+  build_by_entry_and_target('./client/index.jsx', 'public/bundle.js'),
 ]
